Move feed definition to own include and add link real to base template
