+++ torchbox/models.py
@@ -1,9 +1,8 @@
-from datetime import datetime, date, time
+from datetime import date
 
 from django.db import models
 from django.db.models.signals import pre_delete
 from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
-from django.contrib.syndication.views import Feed
 from django.dispatch import receiver
 from django.shortcuts import render
 from django.http import HttpResponse
@@ -22,7 +21,7 @@ from modelcluster.fields import ParentalKey
 from modelcluster.tags import ClusterTaggableManager
 from taggit.models import Tag, TaggedItemBase
 
-from torchbox.utils import export_event, play_filter
+from torchbox.utils import export_event
 
 
 COMMON_PANELS = (
@@ -504,30 +503,6 @@ BlogPage.promote_panels = [
     MultiFieldPanel(COMMON_PANELS, "Common page configuration"),
     ImageChooserPanel('feed_image'),
 ]
-
-
-class BlogFeed(Feed):
-    title = "The Torchbox Blog"
-    link = "/blog/"
-    description = "The latest news and views from Torchbox on the work we do, the web and the wider world"
-
-    def items(self):
-        return play_filter(BlogPage.objects.live().order_by('-date'), 10)
-
-    def item_title(self, item):
-        return item.title
-
-    def item_description(self, item):
-        return item.intro if item.intro else item.body
-
-    def item_link(self, item):
-        return item.full_url
-
-    def item_author_name(self, item):
-        pass
-
-    def item_pubdate(self, item):
-        return datetime.combine(item.date, time())
 
 
 # Jobs index page
