+++ torchbox/templates/torchbox/base.html
@@ -36,11 +36,15 @@
             <link rel="stylesheet" href="{% static "torchbox/css/animate.css" %}">
         {% endcompress %}
 
-        {% if self and self|in_play %}
-            <link rel="stylesheet" href="{% static "torchbox/css/play.css" %}">
+        {% if self %}
+            {% if self|in_play %}
+                <link rel="stylesheet" href="{% static "torchbox/css/play.css" %}">
+            {% else %}
+                <link rel="alternate" type="application/rss+xml" title="Torchbox Blog RSS" href="{% url 'blog_feed' %}" />
+            {% endif %}
         {% endif %}
 
-        <link rel="shortcut icon" href="{% static "torchbox/images/favicon.ico" %}"/>
+        <link rel="shortcut icon" href="{% static "torchbox/images/favicon.ico" %}" />
 
         <!--[if lte IE 8]>
             <script src="http://html5shim.googlecode.com/svn/trunk/html5.js"></script>
